// src/app/models/response/OpeningStockItemResponse.ts
export class OpeningStockItemResponse {
  id?: string;
  openingStockNcId?: string;
  openingStockId?: string;
  supplierId?: string;
  itemId?: string;
  itemName?: string;
  stockQty?: number;
  uom?: string;
  batchNo?: string;
  expireDate?: string;
  expireDateFormatted?: string;
  purchaseRate?: number;
  salesRate?: number;
}