// src/app/models/response/SupplierResponse.ts
export class SupplierResponse {
  id?: string;
  supplierName?: string;
  tin?: string;
  bin?: string;
  tradeLicense?: string;
  tradeLicenseValidTo?: string;
  mobile?: string;
  email?: string;
  address?: string;
  contactPersonName?: string;
  contactPersonMobile?: string;
  supplierTypeEnumKey?: string;
  supplierTypeEnumValue?: string;
  country?: string;
  enabled?: boolean;
  isDeleted?: boolean;
  createdAt?: string;
  createdById?: string;
  lastModifiedAt?: string;
  lastModifiedById?: string | null;
}